tests/justhtml-tests/template_recovery.dat

#data
<template><table><colgroup></colgroup><script></script></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <script>
|   <body>

#data
<template><table><colgroup></colgroup><template></template></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <template>
|             content
|   <body>

#data
<template><table><colgroup><script></script></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <script>
|   <body>

#data
<template><table><colgroup><style></style></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <style>
|   <body>

#data
<template><tbody><colgroup></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|         <colgroup>
|   <body>

#data
<template><tbody></table><script></script></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|         <script>
|   <body>

#data
<template><tr></table><template></template></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|         <template>
|           content
|   <body>

#data
<template><tbody><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|   <body>

#data
<template><tr><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|   <body>

#data
<template><caption>x</caption><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <caption>
|           "x"
|   <body>

#data
<template><title>x</title><tbody></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|   <body>

#data
<template><title>x</title><tr></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|   <body>

#data
<template><title>x</title><td>x</template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|         "x"
|   <body>

#data
<template><optgroup></p></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <optgroup>
|           <p>
|   <body>

#data
<template></p><select></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <select>
|   <body>

#data
<template><tbody><table><tr></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|           <tr>
|   <body>

#data
<template><select><table></select><option></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <select>
|           <option>
|           <table>
|   <body>

#data
<template><base><tbody></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <base>
|   <body>

#data
<template><noframes></noframes><td>x</template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <noframes>
|         "x"
|   <body>

#data
<template></template>&#32;
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|     " "
|   <body>

#data
<body><template></template>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <template>
|       content

#data
<template></br>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag

#document
| <html>
|   <head>
|     <template>
|       content
|   <body>

#data
<template><noscript></noscript>Y
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <noscript>
|         "Y"
|   <body>

#data
<template><dialog><object></dialog>e
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <dialog>
|           <object>
|             "e"
|   <body>

#data
<head><template></template></head><!--c-->
#errors
(1,6): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|   <!-- c -->
|   <body>

#data
<template><caption><tr>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <caption>
|         <tbody>
|           <tr>
|   <body>

#data
<template><xmp></xmp><td>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <xmp>
|   <body>

#data
<template><td></th>x
#errors
(1,10): expected-doctype-but-got-start-tag
(1,15): unexpected-end-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <td>
|           "x"
|   <body>


#data
<template><title>x</title><base><link></template>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|         <base>
|         <link>
|   <body>

#data
<template></html><//
#errors
(1,10): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag
#document
| <html>
|   <head>
|     <template>
|       content
|         <!-- / -->
|   <body>

#data
<template><tr>d
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|         "d"
|   <body>

#data
<!><template></template><?
#errors
(1,1): expected-doctype-but-got-chars
#document
| <!--  -->
| <html>
|   <head>
|     <template>
|       content
|   <body>

#data
<template><d><table><tr>
#errors
(1,10): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|     <template>
|       content
|         <d>
|           <table>
|             <tbody>
|               <tr>
|   <body>

#data
<template><table><tr><table>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <tbody>
|             <tr>
|         <table>
|   <body>

#data
<template><table><col>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|             <col>
|   <body>

#data
<template><colgroup>X
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <colgroup>
|         "X"
|   <body>

#data
<template><colgroup><v>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <colgroup>
|         <v>
|   <body>

#data
<template><col><
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <col>
|   <body>

#data
<template><table><table><
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|         "<"
|         <table>
|   <body>

#data
<template><td><table><tr>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <td>
|           <table>
|             <tbody>
|               <tr>
|   <body>

#data
<template><table><body><td>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <tbody>
|             <tr>
|               <td>
|   <body>

#data
<template><tr><form>V
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|           <form>
|         "V"
|   <body>

#data
<template><tbody><e><tr>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|           <tr>
|         <e>
|   <body>

#data
<template><tr><u><td>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|           <td>
|         <u>
|   <body>

#data
<table><template><n><form><
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <n>
|             <form>
|               "<"

#data
<table><template><tr>s
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <template>
|         content
|           <tr>
|           "s"

#data
<template><tr><script>
#errors
#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|           <script>
|   <body>
